Docs cleanup :books:
