+++ docs/ARC.md
@@ -15,8 +15,8 @@ ARCtrl aims to provide an easy solution to create and manipulate ARCs in memory.
 
 ```fsharp
 // F#
-#r "nuget: FsSpreadsheet.ExcelIO, 4.1.0"
-#r "nuget: ARCtrl, 1.0.0-alpha9"
+#r "nuget: FsSpreadsheet.ExcelIO, 5.0.2"
+#r "nuget: ARCtrl, 1.0.0-beta.8"
 
 open ARCtrl
 
@@ -53,76 +53,27 @@ open FsSpreadsheet.ExcelIO
 let arcRootPath = @"path/where/you/want/the/NewTestARC"
 
 // From ARCtrl.NET
-let fulfillWriteContract basePath (c : Contract) =
-    let ensureDirectory (filePath : string) =
-        let file = new System.IO.FileInfo(filePath);
-        file.Directory.Create()
-    match c.DTO with
-    | Some (DTO.Spreadsheet wb) ->
-        let path = System.IO.Path.Combine(basePath, c.Path)
-        ensureDirectory path
-        FsWorkbook.toFile path (wb :?> FsWorkbook)
-    | Some (DTO.Text t) ->
-        let path = System.IO.Path.Combine(basePath, c.Path)
-        ensureDirectory path
-        System.IO.File.WriteAllText(path,t)
-    | None -> 
-        let path = System.IO.Path.Combine(basePath, c.Path)
-        ensureDirectory path
-        System.IO.File.Create(path).Close()
-    | _ -> 
-        printfn "Contract %s is not an ISA contract" c.Path
-
-// From ARCtrl.NET
 let write (arcPath: string) (arc:ARC) =
     arc.GetWriteContracts()
-    |> Array.iter (fulfillWriteContract arcPath)
+    // `Contracts.fulfillWriteContract` from Contracts.fsx docs
+    |> Array.iter (Contracts.fulfillWriteContract arcPath)
 
 write arcRootPath arc
 ```
 
 ```js
 // JavaScript
-import {ARC} from "@nfdi4plants/arctrl";
-import {Xlsx} from "fsspreadsheet";
-import fs from "fs";
-import path from "path";
-
-const arcRootPath = "C:/Users/Kevin/Desktop/NewTestARCJS"
-
-async function fulfillWriteContract (basePath, contract) {
-    function ensureDirectory (filePath) {
-        let dirPath = path.dirname(filePath)
-        if (!fs.existsSync(dirPath)){
-            fs.mkdirSync(dirPath, { recursive: true });
-        }
-    }
-    const p = path.join(basePath,contract.Path)
-    if (contract.Operation = "CREATE") {
-        if (contract.DTO == undefined) {
-            ensureDirectory(p)
-            fs.writeFileSync(p, "")
-        } else if (contract.DTOType == "ISA_Assay" || contract.DTOType == "ISA_Assay" || contract.DTOType == "ISA_Investigation") {
-            ensureDirectory(p)
-            await Xlsx.toFile(p, contract.DTO)
-            console.log("ISA", p)
-        } else if (contract.DTOType == "PlainText") {
-            ensureDirectory(p)
-            fs.writeFileSync(p, contract.DTO)
-        } else {
-            console.log("Warning: The given contract is not a correct ARC write contract: ", contract)
-        }
-    }
-}
+import {fulfillWriteContract} from "./Contracts.js";
 
 async function write(arcPath, arc)  {
     let contracts = arc.GetWriteContracts()
-    contracts.forEach(async contract => {
+    for (const contract of contracts) {
+        // from Contracts.js docs
         await fulfillWriteContract(arcPath,contract)
-    });
+    };
 }
 
-await write(arcRootPath,arc)
+await write(arcRootPath, arc)
 ```
 
 ## Read
@@ -136,7 +87,6 @@ Setup will be placed on top, with the actual read below.
 open System.IO
 
 // Setup
-
 let normalizePathSeparators (str:string) = str.Replace("\\","/")
 
 let getAllFilePaths (basePath: string) =
@@ -149,26 +99,9 @@ let getAllFilePaths (basePath: string) =
     )
     |> Array.ofSeq
 
-// from ARCtrl.NET
-// https://github.com/nfdi4plants/ARCtrl.NET/blob/ba3d2fabe007d9ca2c8e07b62d02ddc5264306d0/src/ARCtrl.NET/Contract.fs#L7
-let fulfillReadContract basePath (c : Contract) =
-    match c.DTOType with
-    | Some DTOType.ISA_Assay 
-    | Some DTOType.ISA_Investigation 
-    | Some DTOType.ISA_Study ->
-        let path = System.IO.Path.Combine(basePath, c.Path)
-        let wb = FsWorkbook.fromXlsxFile path |> box |> DTO.Spreadsheet
-        {c with DTO = Some wb}
-    | Some DTOType.PlainText ->
-        let path = System.IO.Path.Combine(basePath, c.Path)
-        let text = System.IO.File.ReadAllText(path) |> DTO.Text
-        {c with DTO = Some text}
-    | _ -> 
-        printfn "Contract %s is not an ISA contract" c.Path 
-        c
-
 // put it all together
 let readARC(basePath: string) =
+    // Get all file paths for a given ARC folder
     let allFilePaths = getAllFilePaths basePath
     // Initiates an ARC from FileSystem but no ISA info.
     let arcRead = ARC.fromFilePaths allFilePaths
@@ -176,17 +109,70 @@ let readARC(basePath: string) =
     let readContracts = arcRead.GetReadContracts()
     let fulfilledContracts = 
         readContracts 
-        |> Array.map (fulfillReadContract basePath)
+        // `Contracts.fulfillReadContract` from Contracts.fsx docs
+        |> Array.map (Contracts.fulfillReadContract basePath) 
     arcRead.SetISAFromContracts(fulfilledContracts)
     arcRead 
 
 // execution
-
 readARC arcRootPath
 ```
 
 ```js
+// JavaScript
+import {ARC} from "@nfdi4plants/arctrl";
+import {fulfillWriteContract, fulfillReadContract} from "./Contracts.js";
 
+// Setup
+function normalizePathSeparators (str) {
+  const normalizedPath = path.normalize(str)
+  return normalizedPath.replace(/\\/g, '/');
+}
+
+export function getAllFilePaths(basePath) {
+  const filesList = []
+  function loop (dir) {
+      const files = fs.readdirSync(dir);
+      for (const file of files) {
+          const filePath = path.join(dir, file);
+  
+          if (fs.statSync(filePath).isDirectory()) {
+              // If it's a directory, recursively call the function on that directory
+              loop(filePath);
+          } else {
+              // If it's a file, calculate the relative path and add it to the list
+              const relativePath = path.relative(basePath, filePath);
+              const normalizePath = normalizePathSeparators(relativePath)
+              filesList.push(normalizePath);
+          }
+      }
+  }
+  loop(basePath)
+  return filesList;
+}
+
+// put it all together
+async function read(basePath) {
+    let allFilePaths = getAllFilePaths(basePath)
+    // Initiates an ARC from FileSystem but no ISA info.
+    let arc = ARC.fromFilePaths(allFilePaths)
+    // Read contracts will tell us what we need to read from disc.
+    let readContracts = arc.GetReadContracts()
+    console.log(readContracts)
+    let fcontracts = await Promise.all(
+        readContracts.map(async (contract) => {
+            let content = await fulfillReadContract(basePath, contract)
+            contract.DTO = content
+            return (contract) 
+        })
+    )
+    arc.SetISAFromContracts(fcontracts)
+    return arc
+}
+
+// execution
+
+await read(arcRootPath).then(arc => console.log(arc))
 ```
 
 [1]: <https://www.nuget.org/packages/ARCtrl.NET> "ARCtrl.NET Nuget"
